requirements.txt

blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
docker==7.0.0
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.2
protobuf==4.21.12
requests==2.31.0
urllib3==2.0.6
Werkzeug==2.3.7
